default build path is /build/reproducible-path and switch off log filtering
This switches the default build path to the same that is used by the
Debian buildds, see bug #1034424 and commit
https://salsa.debian.org/dsa-team/mirror/dsa-puppet/-/commit/aed04290e21c4185bb6a2fb3186bfe4cb862198d
This commit also disables log filtering. The previous behaviour can be
restored by adding this to ~/.sbuildrc:

    $build_path = undef;
    $log_filter = 1;

By switching the build path to the same as used by the buildds, uploaded
.buildinfo files can now exactly mimic those that will be created by the
buildds. If the Debian infrastructure does not vary build paths, it does
not make sense to have the behaviour on developer machines be different.

This change will makes it impossible to do concurrent builds of the same
package with the same version using the schroot backend. If you want to
do multiple builds of the same package, consider to either:

 - vary the version for example with --make-binNMU=...
 - manually set the build path via ~/.sbuildrc or --build-path
 - don't use the schroot backend

This change also disables log filtering by default. Log filtering was
introduced in commit baf300b302d2fab45eaa96c06eea2a95b353600d to make it
possible to diff build logs against each other even though their build
path contained a random string. If the build path is not random anymore
by default, log filtering lots its main reason to exist. Disabling it
will make build logs easier to read for newcomers or people outside of
Debian.

At the same time, disabling log filtering will make it harder to diff
build logs against each other for the same package but different
versions. If you need to do this, consider to either:

 - enable log filtering via ~/.sbuildrc
 - process the build log with sed before diffing it