Unshare: preserve HOME environment variable
It has been reported [1] that the unshare backend does not set the HOME
environment variable to /sbuild-nonexistent contrary to other backend.

I tracked the issue to the runuser command, which initializes the HOME,
SHELL, USER, LOGNAME, and PATH environment variables. Fortunately the -p
option allows them to be preserved. However this causes the USER and
LOGNAME environment variables to become undefined. This should not be an
issue for LOGNAME, but is definitely one for USER. To fix that define it
in the call to env, like it is already done for PATH.

[1] https://lists.debian.org/debian-wb-team/2024/05/msg00004.html