configure apt's barbarian architectures for the cross build profile
The cross build profile indicates that we cannot execute host
architecture binaries. It can be used to include additional dependencies
that are needed to avoid running built tools. On the flip side,
Multi-Arch:foreign packages usually contain things we want to run and
that typically means that satisfying them from the host architecture
will not work. While apt prefers satisfying dependencies on
Multi-Arch:foreign packages from the build architecture, it occasionally
does so anyway and when it does things usually go wrong. Configuring the
host architecture as "barbarian" appears to prevent this from happening
in some of the cases.

Addresses: #1071246