lib/Sbuild/Conf.pm: make the schroot binary configurable via ~/.sbuildrc
Moving the SCHROOT option from group __INTERNAL to Programs means, that
it is no longer exempt from the $conf->set() call in
Sbuild::ConfBase::read(). This means, that as part of its CHECK
subroutine, validate_program() will get called and might find out that
schroot is not installed even in situations where the user passed
--chroot-mode=unshare. But command line arguments are only parsed later,
so at the point where the config is parsed, CHROOT_MODE is still set to
its default "schroot". To fix this, make the default of CHROOT_MODE
"undef" and add a GET subroutine which returns the default ('schroot')
if it's not defined. In turn, let the SCHROOT CHECK call $conf->_get
instead of $conf->get, so that it bypasses GET and reads undef.