man/sbuild-update.1: delete clean from an example
The last example in the manual page would call apt-get's update,
dist-upgrade, clean, autoclean and autoremove commands. Since clean
removes all package files from cache, it is unnecessary to call
autoclean thereafter.

In author's humble opinion, autoclean tends to be safer in the long run
(though this may be of less importance in chroots), so instead of
deleting it from the example sbuild-update call, delete clean.