Unshare: run useradd/groupadd with --prefix instead of --root
Calling useradd or groupadd --root does a chroot() into the target
directory, which may not have populated /dev. This can be troublesome
if e.g. the tools try to redirect to redirect to /dev/null, as the
patched src:shadow in Ubuntu does, see [1].

The similar --prefix does not chroot. It has the limits documented in
useradd(8), but they are not relevant in this context. When calling
useradd, we also need to specify --no-log-init, otherwise it will try
to modify the lastlog and faillog databases of the host system.

Similar change in autopkgtest: [2].

[1] https://bugs.launchpad.net/ubuntu/+source/shadow/+bug/2049529
[2] https://salsa.debian.org/ci-team/autopkgtest/-/commit/2344e8197ed29ed