sbuild-usernsexec: avoid getpwnam inside the chroot
getpwnam uses NSS to perform the lookup. It does so after chroot, so we
are running a perl interpreter from one installation inside a chroot of
another. They may differ in distribution major release and architecture
(e.g. when performing a fully qemu-user emulated build). We cannot
expect to be able to load NSS modules. Hence parse /etc/passwd.