lib/Sbuild/Options.pm: replace uses of push by push_conf
Using push may not work as get_conf might return a temporary array from a 'GET' function in Conf.pm.

To handle this case and properly set the conf value with the updated array, use the new `push_conf` function.

`push` usage for EXTERNAL_COMMANDS confs are kept as-is as the conf value is a hash and EXTERNAL_COMMANDS does not have a 'GET' function.