UNSHARE_MMDEBSTRAP_EXTRA_ARGS now supports globs and regexes as keys and is a list instead of a hash
This is an experimental feature. In unshare mode, when mmdebstrap is run
because UNSHARE_MMDEBSTRAP_AUTO_CREATE was set to true, pass these extra
arguments to the mmdebstrap invocation. The option array is given as
key/value pairs. Each key will be matched against strings created from
sbuild configuration variables, namely: DISTRIBUTION,
DISTRIBUTION-BUILD_ARCH, DISTRIBUTION-BUILD_ARCH-HOST_ARCH as well as
against the name of the chroot itself (defined if you use --chroot). If
a key matches one of these strings, the value, containing extra
mmdebstrap arguments is appended to the mmdebstrap argument list. A key
can be a plain string, in which case glob-style expressions can be used.
If the key is a plain string, it has to fully match. If the key is a
plain string, percentage escapes %a and %r will be replaced by host
architecture and distribution of the current build, respectively. A key
can also be a precompiled qr// regular expression but match groups
cannot be referenced in the extra arguments. The value is an array of
extra arguments which are appended to the end of the mmdebstrap exec
array.

Closes: #1088653