Use build start time when generating binNMU timestamp
This ensures that the mtime of all patched files get clamped to the
same timestamp.

When building an NMU, sbuild as run by the buildds first has dpkg apply
patches and *then* will compute which timestamp it puts into
debian/changelog based on the current time. If the former action happens
before a full-second-boundary compared to the latter action (e.g.
patching by dpkg happens at 00:00.900 and gmtime() runs at 00:01.100),
then the modification time of the files patched by dpkg will be *before*
the timestamp that sbuild put in debian/changelog. Thus, those
timestamps will not end up being normalized by clamping them as they are
*before* the timestamp in d/changelog. When a rebuilder then tries to
rebuild the package, the patching process will happen in the future of
the timestamp that was written into debian/changelog, so the
modification timestamp *will* get clamped. This means, that the package
cannot be rebuilt correctly.  For this problem to get fixed, packages on
deb.debian.org which are affected by this problem and have files patched
by dpkg with timestamps before the timestamp in d/changelog have to be
rebuilt.

Bug-Debian: https://bugs.debian.org/1089088
Closes: #1089088