debian/tests/unshare: disable cross-builds until #1088971 is fixed in autopkgtest
This will allow sbuild and packages that trigger it to transition again
even without autopkgtest being fixed and deployed on ci.debian.net.

This will also ease development of sbuild as now the salsa pipeline can
be green again for merge requests even though cross-building is
currently broken due to the gcc situation in #1089007.